# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Arboun - Saudi earnest money management platform
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalDeals: number;
  activeDeals: number;
  completedDeals: number;
  disputedDeals: number;
  cancelledDeals?: number;
  totalAmountEscrowed: number;
  totalAmountCompleted: number;
  pendingSignature: number;
  dealsAsbuyer?: number;
  dealsAsSeller?: number;
}