# lib/api-zod/src/generated/types/deal.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Arboun - Saudi earnest money management platform
 * OpenAPI spec version: 0.1.0
 */
import type { DealStatus } from './dealStatus';
import type { DealType } from './dealType';

export interface Deal {
  id: number;
  title: string;
  type: DealType;
  status: DealStatus;
  amount: number;
  currency: string;
  buyerId: number;
  sellerId: number;
  buyerName: string;
  sellerName: string;
  description: string;
  /** @nullable */
  propertyAddress?: string | null;
  /** @nullable */
  vehicleInfo?: string | null;
  deadline: string;
  platformFee?: number;
  buyerSigned?: boolean;
  sellerSigned?: boolean;
  createdAt: string;
  updatedAt: string;
}