# lib/api-zod/src/generated/types/dispute.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Arboun - Saudi earnest money management platform
 * OpenAPI spec version: 0.1.0
 */
import type { DisputeStatus } from './disputeStatus';

export interface Dispute {
  id: number;
  dealId: number;
  reason: string;
  /** @nullable */
  evidence?: string | null;
  status: DisputeStatus;
  /** @nullable */
  resolution?: string | null;
  openedBy: number;
  /** @nullable */
  resolvedAt?: string | null;
  createdAt: string;
}