# lib/api-zod/src/generated/types/timelineEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Arboun - Saudi earnest money management platform
 * OpenAPI spec version: 0.1.0
 */

export interface TimelineEvent {
  id: number;
  dealId: number;
  event: string;
  description: string;
  /** @nullable */
  actorName?: string | null;
  createdAt: string;
}